Using ActiveRecord::Base.configurations to access database adapter across all versions of Rails 3.0+.
This is done to address a weird edge case: the textacular gem overwrites ActiveRecord::Base#respond_to? and establishes an AR connection. This was unexpected and resulted in our threads creating many Postgres connections. The updated code uses a universal way to access the AR adapter (and is less code).