Install exit handler more consistently
Previously, if the background thread got started by the after_fork hook
of unicorn, or passenger, we'd miss installing the at_exit hook.

We rely more on the exit hook now to persist data during restarts, and
missing it came up in testing.

This change rearranges the place that the at_exit hook is installed, so
that any worker app will get it, even if the background thread is
started later.