Add ability for instruments to ignore any sub-instrumented components
For instance, a future Elasticsearch instrument would instrument a call
to the ES library. But ES (typically) uses Faraday underneath to
coordinate the HTTP call, and one of several HTTP backends to make the
call.

But in the UI of APM, we want all that time allocated to Elasticsearch.
So we turn off children tracing when we start an Elasticsearch call,
ignoring the new libraries deeper in.