Only running the CallSet logic if >= 50 ms has elapsed since the CallSet was initialized.
We really only care about n+1 calls that are slow - this roughly does that by not performing any work unless 50 ms has elapsed since the CallSet was initialized. Note this doesn't mean that the sum of the call time is >=50 ms, just that the time between the first and last call is >= 50 ms. That's easier to check.