Rewrite config file reading to be more resiliant
Now a config instance is made up of several layers:
  * Environment
  * Config file
  * Defaults

During agent boot, the middle layer (the file) just isn't present, so
there aren't issues around attempting to read a file before you can
figure out where it should be and similar.

Error handling is improved, and the file layer won't continually attempt
to re-read a file that is inaccessible.

Also simplifies the API, not having to ask for `env_only` as a safeguard
while reading the Config file.