Move samplers to be pulled from Store, rather than pushed into it.
The underlying motivation is to allow us to line up each minute's worth
of data exactly with the minute of the reporting period. Previously, our
CPU and Memory samplers were off by a period, but it didn't matter since
we weren't attempting to correlate them with changes in the application.
("CPU spike here matched up to slowness there").  The CPU & Memory
samplers were just for a bit of addtional data.

Adding 95th percentile has us matching up data exactly on the charts
now, and the line needs to match exactly to when other events happen.