Fix lookup of configuration values
It was getting the last item in the list that knew a key, rather than
the first. The motivating specifics of this bug was that it was finding
the default dev_trace value (false), rather than the one I specified in
my config file (true).

This bug was introduced in the previous version of this file, when I
added coercions