Adjust BacktraceParser to strip off all leading Rails root, but no more
Given a frame's file:
  /Users/scout/secret-next-big-thing/current/app/controllers/marketing_controller.rb

Before, we were attempting to be left with
`app/controllers/marketing_controller.rb`, and it was explicitly looking
for the `app` component of the path.

Now we are looking only for the application root, allowing lib/* and
similar files to be included as "in-app" backtrace frames.