Remove test made obsolete by previous Layaway refactoring
We need more test around Layaway file, but this is not that test.  Now
that we write many files, we should verify that files go to the right
spots, with the right data.  This test still assumed everything was
going to the same file, which made it hard to adapt.