Protect against errors while writing layaway
Remove the data we attempted to write, even if it failed - prevents a
long-term memory leak if it continually fails.

In the future, we may want a circuit breaker / error-condition checker
that turns off scout_apm if it detects it's in a broken state, but that
is outside the scope of this commit.