Safely load scout_apm.yml for an unknown env
Previously, we were raising an exception letting you know that the
environment we were attempting to find in scout_apm.yml wasn't present.
This caused problems for customers who deployed to "staging", or "testing"
or "QA", and just didn't have a scout_apm config.

Now we log the fact that we couldn't find the environment, and continue
on.