Only call shutdown once per process.
Also adds safety around the call that actually broke if you called
shutdown twice.

The case this happened is in Passenger integration, where we have a hook
for the `stopping_worker_process` event, which called shutdown.  Then
the process actually stopped, which fired our at_exit handler, causing
shutdown to be called a second time.  That second time attempted to work
with the background thread that had been finished in the first call,
causing an error message to be emitted.