Fix incompatibility with Mongo gem 2.1
Mongoid 5 requires Mongo ~> 2.1. In development, we tested against 2.3
(current), but a customer ran into an issue where 2.1 was causing
issues.  After investigation, I found that between Mongo gem version 2.1
and 2.3, the name of the accessor that stores the active filters was
changed. This change is meant to accomodate both newer & older form.