Remove the start_with?('delayed_job') check
It's fine to install the DelayedJob plugin if the gem is installed (so
just check the constant), and then let the plugin start the background
worker if/when we're actually in the worker-side of things.

In the rails-side, the lifecycle / worker stuff never gets called, so
it doesn't matter that this anonymous class is hanging out there.