Add and wireup a LayerChildrenSet to cap the number of children for a layer
This lets us cap the number of children any given layer can accumulate,
then it starts combining them together.

The limit is per type (HTTP, ActiveRecord, View...), and currently is at
1000.   That number is arbitrary, and not based on any research.  Likely
it will change or get larger as we figure out the right balance.

The reason for this is to limit the number of Layer objects we store in
ram over the course of a request. In long running jobs in particular,
the number of layers encountered can get very very large, and even
though they are pretty small objects on their own, the weight of them
can add up.  This approach provides a safety valve to prevent our memory use
going up over a long trace.

Anything going over the limit gets bucketed into a single `HTTP/Merged`
bucket, which shows up as its own line-item in the trace, so overall
times and counts are not lost.