Dynamically create controller instrument module
We need to include these instruments at 2 places in the call stack, and
if you `include` a module that already's been included it appears to be
ignored.

So instead, create a new module with the same content twice, to force it
to work.

The case that this is for is an API controller, to wrap the before/after
callbacks (among other callbacks, but that's the important one).

The wrapping goes top-down, a module nearer to the class being run will
wrap everything included below it. Which is why its not possible to just
wrap the Metal controller. We'd miss all modules between it and the
user's code.

So it starts like:

    YourController
    ActionController::API
    <rails-module>
    <rails-module>
    <rails-module>
    ActionController::Metal
    <rails-modules>

And afterward, we need our instruments immediately after the
ActionController::API line, and the ::Metal line.

    YourController
    ActionController::API
    ScoutApm::Instruments::ActionControllerRails3Rails4Instruments
    <rails-module>
    <rails-module>
    <rails-module>
    ActionController::Metal
    ScoutApm::Instruments::ActionControllerMetalInstruments
    <rails-modules>

Since the Metal and API instruments are almost identical I dry'd up the
code, and had each of the 2 Scout Instruments each include an anonymous
module, so it looks like:

    YourController
    ActionController::API
    ScoutApm::Instruments::ActionControllerRails3Rails4Instruments
    <scout-instrument>
    <rails-module>
    <rails-module>
    <rails-module>
    ActionController::Metal
    ScoutApm::Instruments::ActionControllerMetalInstruments
    <scout-instrument>
    <rails-modules>