Abstract recording approach from TrackedRequest
This allows different approaches on how to handle the TrackedRequest
object after the request is finished.

This PR adds a configuration option to allow tracing of tracked requests
in a different thread than the request itself, allowing it to get out of
the critical path of the response.

With async_recording setting:

```
Before:

Start Req --> End Req --> #record! (& process) --> Store --> Response sent

After:

Start Req --> End Req --> Response sent
                      --> #record! (& process) --> Store
```

The goal is that the time spent recording the details of the request is
now not in the way of returning the request.  The same amount of work is
done, just in a background thread instead of the hot-path of the
request.