Add TracerPrepend to use Prepended Modules
Use that in ActionView to solve an incompatibility with the `jb` gem
which uses Module.prepend on the same methods we instrument, causing a
problem when we used the older style alias_method approach to
instrumentation.

This is not compatible with Ruby 1.8 or 1.9. A dynamic switching or some
other logic is needed to make this work with those older rubies.