Also find queue of a resque job from klass.queue
Also add a fallback. We shouldn't ever allow queue to be nil, even
though I don't think resque will run a job if both @queue and #queue are
missing. Always have a fallback of "unknown", so we can report smething.

This oversight of queue name was manifesting as a call to nil.downcase
when we go to process the JobRecord.