Better naming if a sidekiq `.delay` is called on an instance
I think you can only do this via an extension, but googling made it seem
like that was pretty common to add on.  This gets the underlying name of
the class, instead of the to_s of the instance, which often could be
totally dynamic like "#<Foo:0x007ffd7a9dd8a0>". Ideally, that says
"Foo".

Note that the `Module` type is the superclass of Class.

class Foo
 ...
end

Foo.ancestors #=> Class, Module, ....

f = Foo.new
f.ancestors #=> Foo, .... (no Class or Module)