ensure the background worker is running more aggressively
There are so many forms of app servers, forking, threading, and customer
environments, that it's been a whack-a-mole trying to get the background
worker thread started reliably in every case. So this is a pretty
aggressive fallback. It goes off and checks the background worker.

The Check is a presence & liveness check, so its fast.