Move creation of reporting periods, and samplers to own method
Previously, if a worker got no traffic at all, we would not necessarily
create a ReportingPeriod object for that minute, which in turn would
mean we didn't collect samplers.

The result was memory usage appearing to jump up & down more than it
really was (since we were missing certain processes' readings).

Now, the store has a #tick! method to do any per-minute work, including
creating that reporting period, and reading samplers into it.

ALSO: this fixes another minor issue related to shutdown. When shutting
down, the agent called #write_to_layaway(force). This caused
double-readings of memory when called in the sequence inside of a minute:
  * write_to_layaway normally
  * shutdown
  * write to layaway forced