Ensure instruments are installed correctly
There was logic to avoid installing instruments when an app server
wasn't detected, and also separate logic to start the agent when a web
request came through. The interaction was that it would start the agent,
but not actually install instruments. So an app would report memory &
cpu usage, but not any throughput or traces.

This ensures that instruments are installed more often (when
monitor=true at least)