Add should_load_instruments? to Agent#start
Right now this won't do anything, but it lets us continue to use any
logic we may add to #should_load_instruments? later.

Right now, the logic is: "Yes if dev trace, or Yes if monitor=true".
Since we already check monitor=true at the top of the function, it'll
always be true.