Instruments early via exec_queries in Rails >= 3.2 (#182)
* Instruments early via exec_queries in Rails >= 3.2

Rails 5.2 factored away the `find_with_assocations` method, so a
replacement needs to be found for starting instrumentation without
generating a lot of false noise.

Rails 3.2 introduced a private method `exec_queries` to
`ActiveRecord::Relation` that has survived through to Rails 5.2 and
seems relatively stable at this point. It is a good candidate to achieve
the same instrumentation goals because it is just one call earlier in
the chain of the calls invoked when a query (or group of queries) is
being invoked.

In Rails 3.2 through Rails 5.2, the possible call chains relevant to
instrumentation look like:
* `exec_queries` -> (if eager loading) _uninstrumented methods_ -> (if cache miss) `log`
* `exec_queries` -> (if not eager loading) `find_by_sql` -> (if cache miss) `log`

Rails 3.2 through Rails 5.1 call `find_with_assocations` if eager
loading but it no longer needs to be instrumented because the layer has
been started earlier in `exec_queries`.

In the not eager loading case, both `exec_queries` and `find_by_sql` are
instrumented but the request is intentionally marked `ignore_children`
to avoid starting duplicate layers. The instrumentation in `find_by_sql`
is kept because it is possible for client code to directly call
`Model.find_by_sql("...")` and that should be instrumented as well.

In previous versions of Rails, the original implementation stays as-is,
which looks like:
* (if eager loading) `find_with_associations` -> (if cache miss) `log`
* (if not eager loading) `find_by_sql` -> (if cache miss) `log`