Calculates queue latency correctly for scheduled jobs
If a DelayedJob job is queued for future execution
(`Job.delay.deliver(run_at: 5.minutes.from_now)` or via ActiveJob
`Job.set(wait: 5.minutes).perform_later`), the queue latency should be
calculated from the time the job was expected to run, instead of when
it was created.

DelayedJob sets `run_at` to the current time if no delay is specified,
so this logic should still work correctly for jobs scheduled for
immediate execution.