Use Monitor in Store, and add a locked section
Issue #205 has a situation where the @reporting_periods block was
updated during iteration, causing an exception. I think that the issue
is that we added a new StoreReportingPeriod during write_to_layaway
running. So we were looping and then added a new item, causing the
exception.

This wraps both in correct locks. But that was difficult with a mutex,
since you can't reobtain an already taken mutex. Monitor is the same,
but does allow that reentrant behavior, so this switches over to it.