Instrument Memcached
Adds Memcached instrumentation for apps using the Dalli client.

https://github.com/petergoldstein/dalli

The implementation is strongly based on the Redis one and so far it
has been tested with Dalli 2.7, although it should work fine with
older versions too, since the `perform` method has been the default
chokepoint for instrumentation since the first release.