Toggle install of autoinstruments via a config setting
This allows toggling of the autoinstruments install via a `auto_instrument` config option (default=true).

AutoInstruments are not installed (regardless of the `auto_instrument` config value) if not using a supported Ruby version.

To read the config file, I needed to move the loading of auto instruments in the Railstie after the agent is installed. I'm still seeing the same output after moving this after agent install.