Prevent potential exceptions in PeriodicWork while doing logging
A customer ran into an issue with log_layer_histograms in PeriodicWork
raising an exception (and hitting the rescue in BackgroundWorker,
preventing data from being reported). After some investigation it came
to light that the customer didn't have ActiveSupport in their app (a
background job worker), and so Hash#as_json wasn't being monkey patched
in.

This changes a few things:

* Renames some methods & variables in periodic work to be more clear
  about what everything is for
* Only attempt to log autoinstrument timings if autoinstruments is actually configured
* Changes the as_json call in RequestHistograms to work outside AS (also
  a test)
* devolve `inject` call to use manual summing, so it works on older
  rubies.